+++ CHANGES.md
@@ -1,5 +1,11 @@
 ## 0.8.0-rc.7.dev.2
 
+### UX Improvements
+
+1. Minor improvements for the **DATA SOURCES** panel:
+   * Open Dataset Dialogs titles are now either "Open Local Dataset" or "Open Remote Dataset (via OPeNDAP)"
+   * Download button disabled for the "local" data store
+
 ## 0.8.0-rc.7.dev.1
 
 ### UX Improvements